+++ src/tasty/utils/styles.test.ts
@@ -10,7 +10,6 @@ describe('applyStates', () => {
   function checkAppliance(list) {
     list.forEach((obj, i) => {
       it(`to list ${i}`, () => {
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         expect(applyStates(...obj.input)).toEqual(obj.output);
       });